Bot Documentation

+++ Pacman/Actor.cpp
@@ -217,7 +217,7 @@ bool Actor::canMove(ItskoVector2i& newDirections) const
 	}
 	if (
 		(x + newDirections.getX()) < 0 ||
-		(x + newDirections.getX()) > (int)(map->getMapWidth()) - 1 // x + dx can be equal to width - 1; if the expression is true then there is an error
+		(int)(x + newDirections.getX()) > ((int)map->getMapWidth()) - 1 // x + dx can be equal to width - 1; if the expression is true then there is an error
 		)
 	{
 		return 0; // error, e.g. can't move
@@ -225,7 +225,7 @@ bool Actor::canMove(ItskoVector2i& newDirections) const
 
 	if (
 		(y + newDirections.getY()) < 0 ||
-		(int)(y + newDirections.getY()) > (int)(map->getMapHeight()) - 1 // y + dy can be equal to height - 1, if the expression is true then there is an error
+		(int)(y + newDirections.getY()) > ((int)map->getMapHeight()) - 1 // y + dy can be equal to height - 1, if the expression is true then there is an error
 		)
 	{
 		return 0; // error, e.g. can't move
